+++ app/controllers/items_controller.rb
@@ -27,6 +27,12 @@ class ItemsController < ApplicationController
 
   def edit
     @item = Item.find(params[:id])
+    @item_images = @item.item_images
+    # if @item.user_id == current_user
+    #   render :edit
+    # else
+    #   redirect_to root_path
+    # end
   end
 
   private
