+++ app/javascript/packs/application.js
@@ -12,8 +12,7 @@ require("@rails/activestorage").start()
 require("channels")
 require('jquery')
 require('items_new_price')
-// require('items_new_image')
-require('items_new_image2')
+require('items_new_image')
 
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
