+++ app/views/items/edit.html.haml
@@ -16,14 +16,19 @@
             最大5枚までアップロードできます
           #sell-image
             #previews
+              - @item_images.each_with_index do |img, i|
+                #parent{data: {index: i}}
+                  = image_tag img.url.url, class:"js-img", id: "js-img_#{i}"
+                  .js-remove{"data-index": "#{i}", id: "remove_#{i}"} 削除
+
             .sell-image__box#image-box__container
               =f.fields_for :item_images do |i|
                 .sell-image__box__input{"data-index": "#{i.index}"}
-                  = i.file_field :item_image, type: 'file', name: "item[item_images_attributes][#{i.index}][url]",class:"js-file", id:"item_item_images_attributes_0_url"
+                  = i.file_field :item_image, type: 'file' ,name: "item[item_images_attributes][#{i.index}][url]",class:"js-file", id:"item_item_images_attributes_#{i.index}_url"
                   %label{for: "img-file"}
-                    %i.fas.fa-camera
-                %p.sell-image__box__text
-                  クリックしてファイルをアップロード
+              %i.fas.fa-camera
+              %p.sell-image__box__text
+                クリックしてファイルをアップロード
       .sell-box
         .sell-title__bottom
           .sell-title__group
